put work cache in its own file
